+++ EHR-Sandbox-Api/src/main/java/org/immregistries/ehr/controllers/AuthController.java
@@ -1,15 +1,11 @@
 package org.immregistries.ehr.controllers;
 
-import java.util.HashSet;
-import java.util.List;
-import java.util.Set;
-import java.util.stream.Collectors;
 import javax.validation.Valid;
 
 import org.immregistries.ehr.entities.ImmunizationRegistry;
 import org.immregistries.ehr.entities.User;
-import org.immregistries.ehr.repositories.ImmunizationRegistryRepository;
-import org.immregistries.ehr.repositories.UserRepository;
+import org.immregistries.ehr.entities.repositories.ImmunizationRegistryRepository;
+import org.immregistries.ehr.entities.repositories.UserRepository;
 import org.immregistries.ehr.security.JwtResponse;
 import org.immregistries.ehr.security.JwtUtils;
 import org.immregistries.ehr.security.UserDetailsImpl;
@@ -20,7 +16,6 @@ import org.springframework.security.authentication.UsernamePasswordAuthenticatio
 import org.springframework.security.core.Authentication;
 import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.security.crypto.password.PasswordEncoder;
-import org.springframework.web.bind.annotation.CrossOrigin;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
