+++ EHR-Sandbox-Api/src/main/java/org/immregistries/ehr/controllers/FeedbackController.java
@@ -1,27 +1,13 @@
 package org.immregistries.ehr.controllers;
 
-import ca.uhn.fhir.context.FhirContext;
-import ca.uhn.fhir.parser.IParser;
-import org.immregistries.ehr.EhrApiApplication;
 import org.immregistries.ehr.entities.*;
-import org.immregistries.ehr.logic.PatientHandler;
-import org.immregistries.ehr.logic.RandomGenerator;
-import org.immregistries.ehr.logic.ResourceClient;
-import org.immregistries.ehr.repositories.*;
-import org.immregistries.ehr.security.AuthTokenFilter;
-import org.immregistries.ehr.security.JwtUtils;
+import org.immregistries.ehr.entities.repositories.*;
 import org.immregistries.ehr.security.UserDetailsServiceImpl;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
-import org.springframework.web.server.ResponseStatusException;
-import org.springframework.web.servlet.support.ServletUriComponentsBuilder;
 
-import java.net.URI;
-import java.util.Date;
 import java.util.Optional;
 
 @RestController
