+++ EHR-Sandbox-Api/src/main/java/org/immregistries/ehr/controllers/MainController.java
@@ -9,8 +9,7 @@ import org.immregistries.ehr.entities.Patient;
 import org.immregistries.ehr.entities.VaccinationEvent;
 import org.immregistries.ehr.logic.RandomGenerator;
 import org.immregistries.ehr.logic.ResourceClient;
-import org.immregistries.ehr.repositories.ImmunizationRegistryRepository;
-import org.immregistries.ehr.security.AuthTokenFilter;
+import org.immregistries.ehr.entities.repositories.ImmunizationRegistryRepository;
 import org.immregistries.ehr.security.UserDetailsServiceImpl;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
