+++ EHR-Sandbox-Api/src/main/java/org/immregistries/ehr/controllers/PatientController.java
@@ -3,7 +3,6 @@ package org.immregistries.ehr.controllers;
 import ca.uhn.fhir.context.FhirContext;
 import ca.uhn.fhir.parser.IParser;
 import ca.uhn.fhir.rest.api.MethodOutcome;
-import org.hl7.fhir.instance.model.api.IIdType;
 import org.immregistries.ehr.EhrApiApplication;
 import org.immregistries.ehr.entities.Facility;
 import org.immregistries.ehr.entities.ImmunizationRegistry;
@@ -12,11 +11,10 @@ import org.immregistries.ehr.entities.Tenant;
 import org.immregistries.ehr.logic.PatientHandler;
 import org.immregistries.ehr.logic.RandomGenerator;
 import org.immregistries.ehr.logic.ResourceClient;
-import org.immregistries.ehr.repositories.FacilityRepository;
-import org.immregistries.ehr.repositories.ImmunizationRegistryRepository;
-import org.immregistries.ehr.repositories.PatientRepository;
-import org.immregistries.ehr.repositories.TenantRepository;
-import org.immregistries.ehr.security.AuthTokenFilter;
+import org.immregistries.ehr.entities.repositories.FacilityRepository;
+import org.immregistries.ehr.entities.repositories.ImmunizationRegistryRepository;
+import org.immregistries.ehr.entities.repositories.PatientRepository;
+import org.immregistries.ehr.entities.repositories.TenantRepository;
 import org.immregistries.ehr.security.UserDetailsServiceImpl;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
