+++ EHR-Sandbox-Api/src/main/java/org/immregistries/ehr/controllers/TenantController.java
@@ -2,8 +2,8 @@ package org.immregistries.ehr.controllers;
 
 import org.immregistries.ehr.entities.Patient;
 import org.immregistries.ehr.entities.Tenant;
-import org.immregistries.ehr.repositories.PatientRepository;
-import org.immregistries.ehr.repositories.TenantRepository;
+import org.immregistries.ehr.entities.repositories.PatientRepository;
+import org.immregistries.ehr.entities.repositories.TenantRepository;
 import org.immregistries.ehr.security.UserDetailsServiceImpl;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
