+++ EHR-Sandbox-Api/src/main/java/org/immregistries/ehr/controllers/VaccinationController.java
@@ -7,12 +7,11 @@ import ca.uhn.fhir.rest.api.MethodOutcome;
 import org.hl7.fhir.r4.model.Immunization;
 import org.immregistries.ehr.EhrApiApplication;
 import org.immregistries.ehr.entities.*;
+import org.immregistries.ehr.entities.repositories.*;
 import org.immregistries.ehr.logic.HL7printer;
 import org.immregistries.ehr.logic.ImmunizationHandler;
 import org.immregistries.ehr.logic.RandomGenerator;
 import org.immregistries.ehr.logic.ResourceClient;
-import org.immregistries.ehr.repositories.*;
-import org.immregistries.ehr.security.AuthTokenFilter;
 import org.immregistries.ehr.security.UserDetailsServiceImpl;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
