+++ EHR-Sandbox-Api/src/main/java/org/immregistries/ehr/entities/repositories/TenantRepository.java
@@ -1,4 +1,4 @@
-package org.immregistries.ehr.repositories;
+package org.immregistries.ehr.entities.repositories;
 
 import com.github.javafaker.Bool;
 import org.immregistries.ehr.entities.Tenant;
