+++ EHR-Sandbox-Api/src/main/java/org/immregistries/ehr/entities/repositories/UserRepository.java
@@ -1,4 +1,4 @@
-package org.immregistries.ehr.repositories;
+package org.immregistries.ehr.entities.repositories;
 
 import org.immregistries.ehr.entities.User;
 import org.springframework.data.repository.CrudRepository;
