+++ EHR-Sandbox-Api/src/main/java/org/immregistries/ehr/entities/repositories/VaccinationEventRepository.java
@@ -1,4 +1,4 @@
-package org.immregistries.ehr.repositories;
+package org.immregistries.ehr.entities.repositories;
 
 import org.immregistries.ehr.entities.Patient;
 import org.immregistries.ehr.entities.VaccinationEvent;
