+++ EHR-Sandbox-Api/src/main/java/org/immregistries/ehr/entities/repositories/VaccineRepository.java
@@ -1,4 +1,4 @@
-package org.immregistries.ehr.repositories;
+package org.immregistries.ehr.entities.repositories;
 
 import org.immregistries.ehr.entities.Vaccine;
 import org.springframework.data.repository.CrudRepository;
