+++ EHR-Sandbox-Api/src/main/java/org/immregistries/ehr/logic/ResourceClient.java
@@ -1,16 +1,13 @@
 package org.immregistries.ehr.logic;
 
 import ca.uhn.fhir.context.FhirContext;
-import ca.uhn.fhir.parser.DataFormatException;
 import ca.uhn.fhir.rest.api.MethodOutcome;
 import ca.uhn.fhir.rest.client.api.IGenericClient;
-import ca.uhn.fhir.rest.server.exceptions.ResourceNotFoundException;
 import org.hl7.fhir.instance.model.api.IBaseResource;
-import org.hl7.fhir.instance.model.api.IIdType;
 import org.hl7.fhir.r4.model.IdType;
 import org.hl7.fhir.r4.model.OperationOutcome;
 import org.immregistries.ehr.entities.ImmunizationRegistry;
-import org.immregistries.ehr.repositories.ImmunizationRegistryRepository;
+import org.immregistries.ehr.entities.repositories.ImmunizationRegistryRepository;
 import org.immregistries.ehr.security.UserDetailsServiceImpl;
 import org.springframework.beans.factory.annotation.Autowired;
