+++ EHR-Sandbox-Ui/README.md
@@ -7,7 +7,9 @@ To change the base URI of the API :
 - for production mode edit [src/environments/environment.ts](src/environments/environment.prod.ts)
 
 ### WAR File
-To obtain the war file for this application, execute script `bash compile_war.sh`, which will generate [dist/ehr-sandbox.war](dist/ehr-sandbox.war)
+To obtain the war file for this application, you can execute script `bash compile_war.sh`, which will generate [dist/ehr-sandbox.war](dist/ehr-sandbox.war)
+
+Otherwise executing `mvn clean package` in parent folder will generate a war file with both the ui and api
 
 
 This project was generated with [Angular CLI](https://github.com/angular/angular-cli) version 13.2.5.
