+++ EHR-Sandbox-Ui/package.json
@@ -1,6 +1,6 @@
 {
   "name": "ehr-sandbox-ui",
-  "version": "1.2.0",
+  "version": "1.2.1-Beta",
   "scripts": {
     "ng": "ng",
     "start": "ng serve",
