+++ EHR-Sandbox-Ui/src/app/_components/_forms/patient-form/patient-form.component.html
@@ -49,7 +49,7 @@
                 <div *ngSwitchCase="'select'">
                   <mat-label class="short-label">{{form.title}}</mat-label>
                   <select  *ngIf="form.options; else short" [(ngModel)]="patient[form.attribute]" name="{{form.attribute}}"  aria-label="Font Style" >
-                    <option *ngIf="patient[form.attribute]" selected="{{!patient[form.attribute]}}" value="undefined"></option>
+                    <option *ngIf="patient[form.attribute]" selected="{{patient[form.attribute] == undefined || patient[form.attribute] == ''}}" value="undefined"></option>
                     <option *ngFor="let option of form.options" selected="{{patient[form.attribute] == option.value}}" value="{{option.value}}">{{option.label ?? option.value}}</option>
                   </select>
                   <ng-template #short >
@@ -59,6 +59,27 @@
                 <div *ngSwitchDefault></div>
               </ng-container>
 
+            </mat-card-content>
+          </mat-card>
+        </mat-grid-tile>
+        <mat-grid-tile>
+          <mat-card class="dashboard-card">
+            <mat-card-header>
+              <mat-card-title>
+                Update dates
+              </mat-card-title>
+            </mat-card-header>
+            <mat-card-content class="dashboard-card-content">
+              <mat-form-field  appearance="outline" class="date-field">
+                <mat-label>Creation date</mat-label>
+                <input matInput [matDatepicker]="created" [ngModel]="asDate(patient.createdDate)" disabled>
+                <mat-datepicker #created></mat-datepicker>
+              </mat-form-field>
+              <mat-form-field  appearance="outline" class="date-field">
+                <mat-label>Update date</mat-label>
+                <input matInput [matDatepicker]="updated" [ngModel]="asDate(patient.updatedDate)" disabled>
+                <mat-datepicker #updated></mat-datepicker>
+              </mat-form-field>
             </mat-card-content>
           </mat-card>
         </mat-grid-tile>
