+++ EHR-Sandbox-Ui/src/app/_components/_forms/vaccination-form/vaccination-form.component.html
@@ -84,6 +84,28 @@
             </mat-card-content>
           </mat-card>
         </mat-grid-tile>
+
+        <mat-grid-tile>
+          <mat-card class="dashboard-card">
+            <mat-card-header>
+              <mat-card-title>
+                Update dates
+              </mat-card-title>
+            </mat-card-header>
+            <mat-card-content class="dashboard-card-content">
+              <mat-form-field  appearance="outline" class="date-field">
+                <mat-label>Creation date</mat-label>
+                <input matInput [matDatepicker]="created" [ngModel]="asDate(vaccination.vaccine.createdDate)" disabled>
+                <mat-datepicker #created></mat-datepicker>
+              </mat-form-field>
+              <mat-form-field  appearance="outline" class="date-field">
+                <mat-label>Update date</mat-label>
+                <input matInput [matDatepicker]="updated" [ngModel]="asDate(vaccination.vaccine.updatedDate)" disabled>
+                <mat-datepicker #updated></mat-datepicker>
+              </mat-form-field>
+            </mat-card-content>
+          </mat-card>
+        </mat-grid-tile>
       </mat-grid-list>
     </form>
   </mat-tab>
