+++ README.md
@@ -1,3 +1,7 @@
 # EHR-Sandbox
 The EHR Sandbox is a testing tool developed by AIRA and NIST along with the IIS Sandbox tool.
 It's purpose is to simulate the behaviour of a Electronic health record (EHR).
+
+Execute mvn clean install to generate war file.
+
+Find it in folder "EHR-Sandbox-Api/target/ehr-sandbox.war"
